src/part2/services/orchestrator/requirements.txt
redis>=5.0
openai>=1.0
qdrant-client>=1.9.0
fpdf2>=2.7
